moody-slant-apple-rgb: New function
This is a wrapper around `moody-slant' that manipulates the colors
so that they use the equivalent Apple RGB colors instead of the
regular sRGB colors.  This appears to be necessary because Emacs
on macOS uses a color space for XPM images than for other things.

Usage: (setq moody-slant-function #'moody-slant-apple-rgb).

Re #2.

This is a port of the equivalent hack used by Powerline.
See https://github.com/milkypostman/powerline/issues/54
and https://github.com/milkypostman/powerline/pull/141.