resize-temp-buffer-window: Advice to set mode-line height
When `resize-temp-buffer-window' is called on a newly created window,
then `redisplay' had not been called yet, which means that the
mode-line has not been calculated yet, which means that its height
has not been changed to `moody-mode-line-height' yet, which results
in the bottom halve of the last line being cut off.

Explicitly `redisplay' to prevent that.

The meaning of the argument is different for
`resize-temp-buffer-window' and `redisplay' is different ("the window"
vs. "force redisplay" boolean).  The window is always specified so the
redisplaying is always forced here.