Use old-selected-window or pre-redisplay-functions
Detecting whether the window whose mode-line is currently being
formatted is the actual selected window was believed to be difficult
because that window was always temporarily selected while doing so.

I brought this up on emacs-devel and these two approaches were
suggested.  We can't just always use `old-selected-window' because
that did not exist before Emacs 27.