moody-wrap: Resist mode-line-compact
Previously a space was used as the text underlying an image, which
meant that it was located inside of a span of multiple spaces. That
was perfectly okay until Emacs 28 added `mode-line-compact', which
substitutes just one space for any span of spaces.

To prevent that mechanism from removing our images, we now we use
"|" as the text underlying an image.  That text is never displayed,
so we can use whatever we want; i.e. something other than a space.

We could remove the inner whitespace padding around the images but
that would look really bad, so we don't.  We have no control over
the outer padding if any.

Closes #28.