Avoid calling format-mode-line
Likely due to a regression in Emacs 28 temporarily selecting another
window leads to undesired behavior.  In this case the result was that
after switching to another frame and then back to the original frame
while the minibuffer is active failed to select the minibuffer window
again.  The trigger was that `format-mode-line' temporarily selects
another window.  Using that function was unnecessary and wasteful
to begin with, so this is a win all around.