Improve how the active window is detected
The "active" window (i.e. the one whose mode-line uses the `mode-line'
face instead of `mode-line-inactive') is the selected window or, if
that is the minibuffer window, then the window that was selected
before that minibuffer window became the selected window.

In accordance with the face we also have to use different images whose
colors match those of the face and ideally we would do so by querying
the used face or use the same logic that was used to make the face
determination.  Sadly not all of this information is available from
lisp.

The C macro CURRENT_MODE_LINE_FACE_ID_3 is used to determine the face.
It uses some C functions that have corresponding lisp functions but
some of the information it consults is not fully available from lisp.

Never-the-less we tried to copy the logic in `moody-window-active-p'.
This didn't quite work when switching to another frame while the
minibuffer is active (likely because of differences between
`minibuf_window' and `minibuffer-window').

Additionally, and likely due to a regression in Emacs 28, our use of
`with-selected-window' causes the cursor to jump to the beginning of
the minibuffer when using `minibuffer-complete' with the standard
completion mechanism.  See #29 and previous commit.

We continue to try to use the same logic as the code that determines
the face, but now we make all determinations up-front without trying
to partially use information recorded and exposed by C code.  We do
this using the `pre-redisplay-functions' hook.  This fixes both
issues.