Never draw a line below any tab anywhere
Previously we drew a line below tabs whose left slop faced upward,
except if it belongs to a window at the bottom, right above the
minibuffer.  We did that because information in the mode-line usually
is about the window it belongs to, i.e. the window right above, while
the look of such tabs without a separator lines visually imply
otherwise.  The except was a, probably misguided, compromise to make
things prettier.

This affects our wrapper for `vc-mode', which we change to use a
ribbon instead, to avoid the misguiding visual appearance.

IMO it is best to only use a single tab (for the buffer
identification) whose left slop faces downward, and to reserve tabs
whose left slop face upward for information that is displayed only
temporally in windows above the minibuffer and which is about the
contents of the minibuffer, such as the eldoc message that is
displayed while `eval-expression' reads user input in the minibuffer.