Support wrapping the message ElDoc displays during eval-expression
`moody-replace-eldoc-minibuffer-message-function' and its support
functions are more complicated than other replace functions because
ElDoc hooks into `mode-line-format' in an unusual fashion.

By default ElDoc displays the information in all mode-lines, we only
display it right above the minibuffer where it belongs because it is
about the text in the minibuffer.