Avoid misinterpreting % in buffer names as %-constructs
The names of buffers may contain %, so it is not safe to use them as-is.
We also cannot just replace the call to `buffer-name' with "%b", because
then that string, "%b", not the actual buffer, is used when determining
how much padding has to be added to make the final result at least 20
characters wide.

Instead we call (format-mode-line "%b") in our code.  Beside replacing
"%b" with the name of the buffer, this also sets the `mode-line-elt-no'
property to 1.  I could not find any documentation for that, but one of
its effects appears to be, that % won't be interpreted as a %-construct
further down the line.